A proper version of the form. SPA
The whole form is in single page. The backend simply does verification.
No state maintenance of any sort. The client fills the full form, and
then submits it. In case the server disagrees, the form must be refilled
(trying to make it such that the data filled by the user persists -
LocalStorage).